requirements.txt
blake3==0.3.3
Django==4.2.1
elasticsearch==8.7.0
elasticsearch_dsl
ffmpeg==1.4
imageio==2.28.0
matplotlib==3.7.1
opencv_python==4.7.0.72
Pillow==9.5.0
cryptography==39.0.1
PyJWT==2.6.0
requests==2.28.1
toml==0.10.2
coverage==7.2.2
uwsgi
psycopg2
PyEmail
django-cors-headers
cos-python-sdk-v5
ffmpeg-python